Create Delete Parameter
Delete any specific parameter defined in the list. Examples in this list are:

Dim MyArrayList As New ArrayList
MyArrayList.Add("G_L") <- Specified parameter #1 
MyArrayList.Add("G_") <- Specified parameter #2

Add as many parameters as required. Useful if you are unable to remove a parameter manually, or if one appears to be hidden.