Create Filename Parameter Creation
This code creates a specified parameter based on the file name. Useful for automating emboss details, laser cutting text, etc in a part.

Save as an internal rule on the individual file you wish to drive it from.

Handy when coupled with the automated "File Save As List" rule, for things like name tags etc.